requirements.txt
scapy
python-nmap
tk
